check genesis file exists before creating

+++ node-runner-cli/config/Genesis.py
@@ -1,3 +1,6 @@
+import os.path
+
+
 class GenesisConfig:
 
     @staticmethod
@@ -12,7 +15,8 @@ class GenesisConfig:
 
     @staticmethod
     def create_genesis_file(genesis_json_location: str, genesis: str):
-        f = open(genesis_json_location, "w")
-        filecontent = f"{{\"genesis\": \"{genesis}\"}}"
-        f.write(filecontent)
-        f.close()
+        if not os.path.exists(genesis_json_location):
+            f = open(genesis_json_location, "w")
+            filecontent = f"{{\"genesis\": \"{genesis}\"}}"
+            f.write(filecontent)
+            f.close()
